+++ scripts/new-text-file.ps1
@@ -17,6 +17,8 @@
 param([string]$path = "README.txt")
 
 try {
+	if (Test-Path "$path" -pathType leaf)) { throw "File '$path' is already existing" }
+
 	$pathToTemplate = Resolve-Path "$PSScriptRoot/../data/templates/New.txt" 
 	Copy-Item $pathToTemplate "$path"
 	if ($lastExitCode -ne 0) { throw "Can't copy template to: $path" }
